+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: paula <paula@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/28 17:46:41 by paula             #+#    #+#             */
-/*   Updated: 2023/03/01 15:15:32 by paula            ###   ########.fr       */
+/*   Updated: 2023/03/02 10:10:25 by paula            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int     ft_isdigit(int c);
 int     ft_isalnum(int c);
 int     ft_isascii(int c);
 int     ft_isprint(int c);
-int     ft_strlen(const char *str);
+size_t     ft_strlen(const char *str);
 void    *ft_memset(void *s, int c, size_t n);
 void    ft_bzero(void *s, size_t n);
 void    *ft_memcpy(void *dest, const void *src, size_t n);
